✨ Add quit command
Ideally, the program will accept commands without a trailing newline, but that looks to be very difficult in Elixir/Erlang.  So for now, this requires the user to hit `Enter` to run a command.

One possible solution is to use `ex_ncurses`, but that might cause some portability problems. However, we might need it for later features so we'll consider it at that time.