🐛 Fix issues with listener loop
- Using `IO.getn` was resulting in double test runs.

- We were re-running tests on unrecognized commands.

- `Ctrl-D` (EOF) wasn't being handled properly

Switched to using `IO.gets` to read commands.

Extracted a CommandProcessor for handling the commands; it handles `:eof` the same way as the `q` command, trims whitespace from string commands, and ignores unknown commands.

The CommandProcessor takes and returns a config; it currently doesn't make any changes to it, but future commands will.

The main loop now only runs the tests when a non-quit command is recognized.