🔥 Remove some config options
Because `mix test.interactive` is geared towards running `mix test` interactively, some of the configuration options inherited from mix-test.watch don't make as much sense.

With interactive mode, changing the CLI executable from `mix` to `iex -S mix` doesn't work, so we now hard-code that to `mix`.  If there is another use case for this option, I'm willing to reconsider.

The support for running multiple tasks was not ideal, as it would pass the same arguments to all of the commands, which is probably not what we really want (see https://github.com/lpil/mix-test.watch/issues/80) for example.

However, there is still a use for defining a task alias that runs `mix test` under the hood, so we still allow configuration of a single task.  However, `mix test.interactive` assumes that the custom task accepts the same command-line arguments as `mix test`.

Again, if there is a legitimate use cas e for running multiple tasks, I'm willing to consider adding that feature back, but I'd want to find a way to allow different arguments to be passed to different tasks.