✨ Initialize patterns from command-line
Initializes the matching patterns from the command-line.  Patterns are applied as if they were specified by the `p` command.

Ignores any filenames provided by the `--include` or `--only` flags.

In order for this to work properly, we need to know all of the options and their types supported by `mix test` in order to properly parse the command line.  Otherwise, a user could run `mix test.interactive --trace pattern` and the option parser would treat `--trace` as a string with the value `pattern` rather than as a boolean flag with a separate `pattern`.

The downside to this is that it introduces a compatibility issue if a future version of `mix test` adds a new command-line option. Until we update our list, users will not be able to pass the new option through `mix test.interactive`.

At present, the tradeoff feels worth it, because now users can start `mix test.interactive` with an initial pattern without having to wait for the initial test run.