♻️ Move command handling into GenServer
With all command-handling and test running in the same GenServer, commands are now run one at a time.

It is no longer possible to run a command while a test run is happening due to file changes. Instead, any commands that come in while the tests are running will be processed once the test run completes.