🐛 Avoid timeout errors
When a command is entered during a long test run, a GenServer timeout would be raised.

To fix this, `InteractiveMode.process_command` now uses `cast` instead of `call` so that the read loop doesn't block on commands.

To make this work, we now have to use `System.stop` for the `quit` command rather than returning a `:quit` atom and allowing the loop to exit naturally.

`System.stop` takes just enough time that it seems like the quit command was ignored, so I also added a `Shutting down...` message to make it clear that something is happening.