🐛 Fix `p` command for umbrella projects
In an umbrella project, the `p` (pattern) command was not matching any files.

This happened because it uses `test` as the default path to find files to match, which doesn't exist in an umbrella project.

With this change, we detect an umbrella project and in that case, wildcard-match the `test` directories of all of the apps in the umbrella.

It is still possible to specify the `:test_paths` config option for `mix test`. If that config setting is present, `mix test.interactive` will use that setting and not default to the wildcard setting.