Update ignored dirs
Added the .elixir_ls and .git dirs to the list of ignored directories in
the Paths module. Otherwise I was getting duplicate runs due to files in
the .elixir_ls directory being changed too often during the execution of
the tests. I'm not sure *why* files in that directory are getting changed
but they are, and this prevents the double test run.