Reverse logic for determining if event is trigger
Because the list of events that should trigger a test run is potentially
much smaller than the list of all possible filesystem events,
particularly on platforms such as MacOS, it is much simpler to specify
the list of events we *do* want to process rather than the list of
events we want to ignore.

N.B. the `FileSystem` module does not completely normalize the names of
the events between platforms, so our list of trigger events includes a
few items that look similar in order to support all platforms.