Ignore some filesystem events emitted by inotify (#86)
The :isdir, :closed, and :attribute events likely don't need to be
monitored to run tests, and at least the :attribute event is causing
problems for projects that use the Domo library for validation. This is
because the Domo library causes a delayed file read event to occur on
some elixir files in the lib directory, which causes their atime
attribute to be updated. This is happening after the test run kicks off,
and then it causes a second test run to occur.

Because the list of events that should trigger a test run is potentially
much smaller than the list of all possible filesystem events,
particularly on platforms such as MacOS, it is much simpler to specify
the list of events we *do* want to process rather than the list of
events we want to ignore.

N.B. the `FileSystem` module does not completely normalize the names of
the events between platforms, so our list of trigger events includes a
few items that look similar in order to support all platforms.