✨ Allow passing mti config as cli args
So far, only the `--clear`/`--no-clear` option is supported; the rest will come in a later commit.

Config options must come first and be separated from `mix test` options by a `--` separator.

The pre-existing `--watch`/`--no-watch` flag can appear in either group for backwards-compatibility, but if it is specified after the `--` separator, a deprecation notice will be printed.

OptionParser has built-in support for recognizing the `--` separator, but I was unable to take full advantage of that so that I could both preserve existing behavior and allow omitting the `--` separator when all arguments are intended for `mix test`.