💥 Configless operation (#108)
* 🏗️ Process command-line args before starting application

In order to prepare for config-less operation, we need to process the command-line args before starting the main parts of the application.

In order to make that work well, we only start a `DynamicSupervisor` intiially and then start the real supervisor later once we've processed the command-line arguments.

* ♻️ Extract argument parsing from Settings -> CommandLineParser

This sets the stage for specifying config options on the command line in addition to settings.

* ♻️ Return config from CommandLineParser

Sets the stage for allowing config settings to be passed on the command line.

* ♻️ Clean up Config initialization

- Rename new -> load_from_environment to make what's happening more explicit
- use a raw `%Config{}` struct where env loading isn't needed
- Eliminate duplicate specification of config defaults
- Inline AppConfig.fetch - separate module for this is overkill right now

* ✨ Allow passing mti config as cli args

So far, only the `--clear`/`--no-clear` option is supported; the rest will come in a later commit.

Config options must come first and be separated from `mix test` options by a `--` separator.

The pre-existing `--watch`/`--no-watch` flag can appear in either group for backwards-compatibility, but if it is specified after the `--` separator, a deprecation notice will be printed.

OptionParser has built-in support for recognizing the `--` separator, but I was unable to take full advantage of that so that I could both preserve existing behavior and allow omitting the `--` separator when all arguments are intended for `mix test`.

* 💥 Add remaining config options

This includes adding an `--exclude` option that now conflicts with mix test's option of the same name.

I also realized that I've broken backwards-compatibility with the `watch`/`no-watch` option.

I'll decide whether this feature warrants a major version bump and then either fix backwards-compatibility or remove the (not quite correct) handling of the legacy watch/no-watch option.

* 👽 Add new mix test options from latest Elixir

* 💥 Remove backwards-compatibility for `--watch`

We've decided to make config-less operation a breaking change. Adjusting to it isn't terribly difficult and simplifies the code and maintenance burden.

* 🥅 Return errors on invalid mti options

Refactors CommandLineParser to return `:ok`/`:error` tuples.

* 🥅 Convert raised exceptions into error responses

* 🏷️ Fix typespec on parse

* ➖ Remove locked temporary_env dep

* 🚸 Improve formatting of usage message

* 🥅 Wrap all errors in a UsageError

* ♻️ Refactor mti argument parsing

- After calling OptionParser, do a pass to parse special values from strings (regexes and runner modules). This allows more helpful error messages
- Then do a pass to combine `:keep`-style arguments into a single list to simplify config building

* ♻️ Don't rely on exceptions for control flow

Modify the option value parsing logic to return :ok/:error tuples and use reduce_while to collect the results.

There are still two `rescue` clauses, but they're used in order to nicely format error messages since formatting functions aren't exposed by `OptionParser` and `Regex`.

* ✨ Add --help option

Only show usage information when requested. When there is an error, display a note about using --help instead.

* ✨ Add a --version option

* 📝 Update documentation