✨ Make ANSI output configurable
Previously, we always enabled ANSI output on Unix-like systems and we didn't on Windows.

However, since Windows 10, ANSI has been available on Windows as well (if the [appropriate registry key is set](https://hexdocs.pm/elixir/IO.ANSI.html)), so we provide a way for users to opt-in if desired.

It is also now possible to run ANSI output off on Unix-like systems if necessary for any reason.

ANSI output can be enabled/disabled via the `ansi_enabled` configuration setting or the `(no-)ansi-enabled` command line option.

We also modify `PortRunner` to take the `os_type` from ProcessTree instead of passing an argument, which matches the changes we made to Config to support the new setting.

We significantly refactor the PortRunner tests to eliminate many near-duplicate tests.