🚸 Improve ansi-enablement on non-Windows systems (#132)
mix_test_interactive automatically enables ANSI output when running on non-Windows systems.

However, the way we were doing it caused the client application to be started when enabling ANSI output and before running the test task.

Rather than using `run -e` to enable ANSI, we now use `eval` instead, which doesn't start the client application.

As part of this work, we also fix a bug where we weren't handling the `--no-start` flag correctly. Previously, we were grabbing that flag (if present) and passing it to the `run` task, but not to the subsequent test task. As a result, the application would still be started in the test task, which is not what the caller would expect.