👽 Address forthcoming deprecation in Elixir 1.20
ExUnit in Elixir 1.20 deprecates `ExUnit.Filters.parse_path/1` in favor of `ExUnit.Filters.parse_paths/1`. However, the newer function is only available in Elixir 1.16 and newer and we (and Elixir) still support 1.14 and 1.15.

To address this, we conditionally use the newer function in 1.20 and above (including pre-releases) and the older function otherwise.