✨ Allow relative pathname patterns
`mix test` allows filenames of the form `./test/hello/hello_test.exs`, but `mix test.interactive` did not.

While we treat command-line arguments as patterns and not filenames, it's confusing that something that works with `mix test` doesn't work with `mix test.interactive`.

We now convert relative pathnames into patterns that will match the list of test files we find in the project.

While `mix test` also supports absolute paths and paths outside of the current project, we choose not to do that here.

Closes #142