⬆️ Upgrade dependencies (#147)
Update Erlang/Elixir and all dependencies to their latest versions.

Loosen the dependency on process_tree to allow the latest v0.3.0 in addition to previous versions. This helps to avoid version conflicts in client projects. There is some danger in being so flexible in case there are breaking changes, but process_tree is pretty stable and we're using its core functionality which is less likely to change in the future.